fix(matter): fix matter startup (#33)
Server::Init was commented out due to some testing issues with
commissioner and commissionee dual roles. This has been put back in, but
the dual roles wont work until other updates are done.

Also fixed a crash when there was no network credentials provider set.

Started a sample network credentials provider for the reference app that
returns hardcoded data for now. This should be made to return something
meaningful for the runtime environment.