build(matter): remove matter build from Dockerfile (#34)
Changes to Matter happen often enough to make rebuilding the docker
image a poor choice for picking them up. At least for now, we should
build Matter externally to the docker image.

The Dockerfile was also not using build stages properly, so they were
removed for now as they are likely causing caching issues.

Here is a top-level script that does it. It is integrated with the
toplevel build.sh script and will only build matter if it isnt already
built.