build: remove all references to ssh necesseties in the image (#40)
Our initial approach to the build process was to handle the building of
Matter within the image. Building Matter requires access to private
repositories, so the home directory of the user was passed as a build
context to the builder to allow access to ssh credentials. Now that
Matter is built within the container, and the home directory is mounted
into the container at run time, this additional build context is not
necessary.

Remove all logic pertaining to ssh credentials at image build time.