build: move watchdog code to cmake option (#44)
Previously it was expected that if the software watchdog option was
enabled that the client would also need to set a preprocessor value for
what code to use when encountering a zigbee core watchdog failure. This
change makes that value a proper cmake option with a default value that
matches what is expected by the only client that uses it.