fix(build): image name clashes on shared machine (#52)
The way the compose file was set up was to look for a `barton_builder`
image to set up the developer environment. When this image is created,
it also creates a corresponding user for whoever initiated the build.
When several users are using a shared machine, clashes can occur when
one user builds `barton_builder`, then another user tries to use it to
set up their developer environment.

To resolve these clashes, adjust the compose file to look for a unique
image name per user, i.e. `barton_builder_$USER`.