fix(tests): initially imported tests (#50)
Several tests were broken as a result of the migration from Zilker to
Barton. This commit does the following to fix these tests:

- Fix mocked calls to `__wrap_fflush` where logging statements are
calling `fflush` but no value was defined for the mocked call.

- Some tests would make calls to `UpdateReportIterval` where the class's
`reportCeilSecs` member was defaulted to a large default value (> 20000)
during class construction. This would make calls to set the
`reportCeilSecs` to a smaller value fail because the ceiling would be
set to the `max` of the large default value or the smaller value. To
fix, define the `reportCeilSecs` member to 0 at class construction time.

- Two tests were not able to be found during run time, so as a result
would fail. Zilker explicity 'excluded' these tests from the build
process which is why they were not able to be found. This exclusion is
not necessary in Barton, remove it.

- Compilation of `libdeviceDescriptorsTest.c` failed after including the
test. Move the inclusion of `setjmp.h` to occur before including
`cmocka.h` as stated in cmocka.h to fix.