build(openssl): include downgraded 1.1.1v version (#70)
Using OpenSSL 3.0 in our docker image causes a bunch of deprecated
function warnings to spew forth when building. We still want to keep
OpenSSL 3.0 because Ubuntu 24.04 does not support 1.1.1, but this change
installs OpenSSL 1.1.1v in addition to 3.0 and uses 1.1.1v specifically
for building Matter and Barton. This also requires the installation of a
libcurl that's linked against OpenSSL 1.1.1v.

Refs: BARTON-147