revert: build(openssl): include downgraded 1.1.1v version (#70) (#77)
This reverts commit 4f8348d75bd160758ffa7a4278383a0a009c71b4.

CI was made and tested on main before the downgrade to openssl 1.1.1 was
introduced. Shortly after the downgrade to openssl 1.1.1 was merged, the
CI commit was merged, and since then is reporting test failures around
finding openssl 1.1.1. To get CI functioning again, this commit reverts
the downgrade to openssl 1.1.1 so we can get CI functioning with this
downgrade in isolation, then get the downgrade merged back into main.