feat(api): include property name when not found (#85)
Several property getters would report a warning when a property couldn't
be found, but wouldn't include the name of the property. This would
result in somewhat useless logs since you can't determine what the
problematic property is.

To improve logging around these warnings, include the property name in
the warning message.