build(matter): Build Barton Reference Matter Library (#100)
This commit introduces the necessary pieces needed to build a matter sdk
static library for barton from the upstream matter sdk rather than a
private fork. It adopts a similar process as example apps use and is
meant to be both a supplemental library for barton development and a
reference for clients to use when trying to create a matter sdk library
for barton.

In addition, this commit:
	1. Refactors some dependency areas such as setting BDS_MATTER_LIB
	   to BartonMatter and cleaning up some headers includes that
	   aren't appropriate.
	2. Fixes the matter sdk on the v1.4.0.0 tag from upstream - not
	   longer using the comcast fork.
	3. Pushes asan/ssp for matter into lower levels.
	4. Deletes references to rdkbcluster and chime.
	5. Updates documentation.
	6. Modified dockerw to use a tmpfs supporting exec mount
	   property.

Refs: BARTON-167