fix(matter): Shutdown chip::Server (#124)
This commit adds an explicit chip::Server shutdown. If not done, bad
things can happen around memory lifecycle. In particular, we are using a
CommonCaseDeviceServerInitParams instance (which we should not be using,
it's specifically for matter repo examples) which supplies static
duration memory to chip::Server. But Server is also static duration and
so you can run into problems in which dtors are called in an undefined
order and result in a use-after-free. According to documentation, we
should be shutting down chip::Server explicitly and not relying on its
destructor to do so.

Refs: BARTON-152