build: isolate user creation (#147)
In order to have a custom user in a devcontainer, VSC requires this user
already exist in the image before the devcontainer starts. Our initial
approach to handling this was for every user to have their own Barton
image, like `barton_builder_$USER`. While this worked, having one base
image that each user can implement is a better approach.

In order to accomplish this goal:

- Create the user at run-time when in the CLI environment by expanding
  the `entrypoint.sh` script.
- Expand on the extended `compose.devcontainer.yaml` to override the
  base image and create a VSC specific image that itself extends the
  base `barton_builder` image. This will ensure the image has the user
  baked in as devcontainers require, but still implement the base
  `barton_builder` image.
- Rename uses of `barton_builder` to full registry name
  `ghcr.io/rdkcentral/barton_builder` to facilitate storing image in
  the GitHub Container Registry.