fix: get device metadata failure (#164)
Attempting to gather metadata details was failing because the
metadata was being stored incorrectly. This occurred because previous
changes in this area allowed the `endpointId` within `setMetadata` to be
an empty string when previously it would expect NULL if the metadata was
on device.

Add logic to ensure `endpointId` is passed as NULL to `setMetadata` when
the metadata is on device to rectify.