feat(thread): Update to latest otbr (#168)
Update to latest ot-br-posix.  This required updating our existing
patches, installing the openthread include directory, adding
openssl to ld path.

We were also passing in the host environment dbus socket, which
not only caused problems for starting up otbr-agent, but would
not work right when on a shared development machine.

The script/start-simulated-otbr.sh works and can be run after
starting the devcontainer.

Refs: BARTON-216, BARTON-217