test: Initial Mock ZHAL Implementation (#165)
This commit lightly restructures the organization of our existing test
harness code to carve out a section for mocks and testing for mocks. It
also adds the very initial parts of a mock zhal implementation to
support zigbee tests. Much more will be needed. Some basic unit testing
for these new mocks are supplied but more is necessary.

Refs: BARTON-205