build: use git tags to determine current image tag (#178)
There a few locations where the default image tag is hardcoded. This can
lead to complications and relies on developers to remember to update it
when a newer docker image is made. Rather than hardcode strings, the
current docker image should be tagged in the git history and parsed when
needed.

- Update `setupDockerEnv.sh` and `action-constants.yaml` to dynamically
  determine the docker image tag based on the most recent
 `docker-builder` git tag reachable by `HEAD`.