fix: Handle LoadOrGenerateLocalNodeId() failure (#184)
Matter::LoadOrGenerateLocalNodeId() can fail and return an invalid Node
ID, but this is not accounted for currently. This fix checks for and
reports this failure. Matter initialization will also fail if this
fails.
In addition, the logic for generating the Node ID is meant to try 10
times to get a valid one, but it wouldn't actually return upon
generating a valid Node ID, meaning it would randomly generate Node IDs
10 times no matter what. This has been fixed.

Refs: BARTON-199