feat!: CommissionableDataProvider implementation to replace test one (#183)
Replace the TestOnlyCommissionableDataProvider with a real
implementation.

The TestOnlyCommissionableDataProvider uses some of the same values that
the Matter sample apps use, so Barton should have its own
implementation.

As per Matter 1.4 spec 5.1.7., Matter can support a dynamically
generated setup passcode. The spake2+ salt is used ephemerally when
establishing a PASE session and can also be randomly generated. The
spake2+ verifier is then generated using the aforementioned passcode and
salt (and an iteration count of our choosing). The generation of all
these fields is handled during the mandatory initialization of our
custom CommissionableDataProvider. The discriminator value is passed on
to the CommissionableDataProvider during Matter initialization. This
implementation leverages Barton's property provider interface to store
all of these values as properties. If a client wishes to pass along the other
values to the CommissionableDataProvider (salt, passcode, etc), they can
do so through the property provider, and those values, if valid, will be
used over the generated ones.

BREAKING CHANGE: Clients consuming Barton MUST set the
`B_DEVICE_SERVICE_BARTON_MATTER_SETUP_DISCRIMINATOR` property with a
discriminator value in production builds OR enable the `BDS_DEV_BUILD`
option for dev builds such that a hardcoded default value can be used.
Matter will fail to initialize if this is not done.
 
Refs: BARTON-198