feat(matter): Add CertifierDACProvider (#203)
Much like 2d20fe0, this commit expands on the concept of matter
providers with the very first provider implementation being a first
pass at the libcertifier DAC provider. This class simply takes a path
to a pkcs12 file (and optional password) that contains the DAC chain
and associated key information.

For development, Barton will continue to use the chip Example DAC Provider.

Refs: BARTON-144