fix: Fix LSAN suppressions to detect leaks again
We were overzealously suppressing glib leaks which backfired on our own
ability to detect leaks with ASAN-enabled compilations. This commit
refactors the LSAN suppression file to include glib again, suppressing
surgically what things can pop up in our integration tests that are
outside our control.

This commit also fixes a leak in recent CommissionableDataProviderTest
that was exposed as part of the suppressions fix.