build!: Refactor library name
This commit refactors the library targets of barton to BartonCore and
BartonCoreStatic. Additionally, we don't attempt to install
BartonCoreStatic as it doesn't make sense to install both the shared
object and the static library. Clients relying on the static library
should reference the cmake target directly or install it out of band.

In addition, this commit removes "device service" naming from public
interfaces and cmake variables. The goal here is to standardize on the
"barton core" nomenclature. Internal modules will still use "device
service" if they already were.