build!: Refactor library name
This commit refactors the library targets of barton to bDeviceService
(and bDeviceServiceStatic). Additionally, we don't attempt to install
bDeviceServiceStatic as it doesn't make sense to install both the shared
object and the static library. Clients relying on the static library
should reference the cmake target directly or install it out of band.