build: Downgrade required cmake to 3.16.5
In order to facillitate some yocto environments that are still on cmake
version 3.16.5, this commit downgrades the required project version.
Ideally, we should raise it back up in the future when environments
update.

This commit also gates the dependency check for libcertifier behind
BDS_MATTER since that library is only needed to support Matter at the
moment.