fix(matter): Fix possible null dereference
If a client compiled with the default CertifierDACProvider but did not
properly set the required properties, the class would default passing
nullptr to std::string's constructor. This is undefined behavior and
would often lead to a null dereference and a crash.

While a client SHOULD set "device.matter.dacP12Path" and
"device.matter.dacP12Password" to properly use the CertifierDACProvider,
failure to do so should not crash the program.

Refs: BARTON-241