feat(matter): Client provided commissioning params (#32)
This change unifies the discriminator and passcode used for
commissioning between remote requests to open commissioning and
internal/client invoked through the Barton API. In both cases the
registered CommissionableDataProvider is used to provide these
parameters.

This change also has the reference app set the discriminator and
passcode to the same values as the ones used commonly in the Matter SDK
if they are not already set.

Refs: BARTON-260

---------

Co-authored-by: cleithner-comcast <87389808+cleithner-comcast@users.noreply.github.com>