feat(matter): Allow xpki host configuration
Previously, we had an internal property, "device.matter.environment" to
configure which certifier host to use. This was particular to comcast and
only supported stage/prod. This change introduces new public property
"barton.matter.operationalHost" which allows clients to set which xpki
certifier host to use. The api mechanism is now deprecated, but we will
fall back to it if this new property is not set.