build: Add BCoreMatterHelper
This commit provides a new cmake module to better interface
to select and query for Matter providers and delegates. Currently,
this implementations sits on top of the existing public options.
This is because these options must exist still to allow clients
a mechanism to select these providers/delegates via command-line
arguments. As a result, the implementation of this module modifies
these cache variables forcibly. For most places (including initial
cache files), this is fine and you can freely use the module.
For options.cmake, it is still required to define the defaults
using bds_string_option for transparancy of defaults as well as
for correct options dump in the cmake output. In addition, you
would not want these defaults to force over client selected
ones.