feat(matter): Upgrade to using Matter 1.4.2.0
This commit updates BartonCore to use Matter v1.4.2.0. This is a
lock-step change. Clients need to update their Matter version in order
to use this SHA / impending release. This is not classified as a
breaking change since the API layer of libBartonCore is not changing.

Clients are encouraged to define a higher-than-default count for
INET_CONFIG_NUM_TCP_ENDPOINTS and INET_CONFIG_NUM_UDP_ENDPOINTS in their
matter project config header (we have defined ours to 50). The number
depends on the expected number of runtime inet addresses that can be
enumerated (both ipv4 and ipv6) from interfaces. An overflow of this cap
can manifest in strange VerifyOrDie "not Constructed" type errors (the
real underlying error was a silent failure to construct).

In addition to compatibility fixes, this commit also removes a
superfluous OTA Software Update Provider cluster entry in barton zap.