fix(matter): return software version string
The DeviceInstanceInfoProvider's GetSoftwareVersion method was not
being overridden which resulted in a not implemented error and a
read failure on the attribute in the basic information cluster.
The software version is handled differently than the hardware version
in the Matter SDK since it changes and is used by the OTA classes.
For now, both the numerical and string versions used are provided
at build time.