fix: Fix RegisterResources()
We had two methods for creating and registering resources in the
base Matter device driver, CreateResources() and RegisterResources().
CreateResources() was not being called, and having two methods responsible
for the same thing was redundant, so this change moves all functionality
into RegisterResources(). It also expands the Matter Light integration test
to check for the existence of the common resources that are created
by RegisterResources().