fix(matter): Verify NOC before saving (#111)
We had an incident where our NOC CA had a regression. Our operational
credentials issuer did not catch this regression before giving to "okay"
on the NOC. As a result, we saved the NOC and never requested anther
one. This NOC continuously failed Matter startup.

To remedy this, this commit creates an new method in
BartonOperationalCredentialsDelegate.hpp called
GenerateNOCChainSynchronous which will do a validation on generated
NOCs and is what is now called directly by Matter.cpp as well as
GenerateNOCChain (which is the async path called by the Matter stack
when commissioning devices).

Refs: BARTON-327