feat: Matter power source cluster and battery attributes (#92)
Add support for the Matter power source cluster and track changes in the
battery percentage. Assumes one power source per device for now.

Known Issues:
- Most of the example apps in the Matter SDK (excluding a few like the
all-clusters-app) do not implement the PowerSource cluster (probably for
simplicity's sake). If we were to fail the commissioning process
whenever we failed to initialize a PowerSource cluster, then we'd fail
to commission most of these example apps during testing, so the code was
written so that commissioning proceeds regardless.

Refs: BARTON-305